fresh-install.bash: prettier consoling

+++ fresh-install.bash
@@ -1,3 +1,12 @@
 #!/usr/bin/env bash
 
-cp fresh-install-printer-cfg ~/printer.cfg
+echo "** ${BASH_SOURCE[0]} **"
+
+if [[ ! -f ~/printer.cfg ]]; then
+	echo "copying a basic printer.cfg to user root."
+	cp fresh-install-printer-cfg ~/printer.cfg
+else
+	echo "~/printer.cfg already exists."
+fi
+
+echo "** end ${BASH_SOURCE[0]} **"
